Move detection.hpp to arch/
