+++ include/eve/arch/detection.hpp
